// components/Molecules/InputDatePickerTarjeta/types.ts
import { Dispatch, SetStateAction } from "react";

export type InputDatePickerTarjetaProps = {
  formData: DataToValidate
  setFormData: Dispatch<SetStateAction<DataToValidate>>
  disabled?: boolean
}

export interface DataToValidate {
  card: string;
  cardError: string;
  nombreTitular: string;
  nombreTitularError: string;
  fechaVencimiento: Date,
  fechaVencimientoError: string;
  dni: string;
  dniError: string;
  codigoSeguridad:string;
  codigoSeguridadError: string;
  preferedCuotas: {cuota: string, installment: number};
}